+++ src/main/java/lotto/config/LottoConfig.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public class LottoConfig {
     public static final int LOTTO_PRICE = 1000;
-    public static final List<Integer> awards =
-            List.of(2000000000, 30000000, 1500000, 50000, 5000);
+    public static final List<Long> awards =
+            List.of(2000000000L, 30000000L, 1500000L, 50000L, 5000L);
 
     public static void validate() throws IllegalArgumentException {
         try {
